Single Responsibility Principle
This principle states that "A class should have only one reason to change."

In layman terminology, this means that a class should not be loaded with multiple responsibilities and a single responsibility should not be spread across multiple classes or mixed with other responsibilities. The reason is that more changes requested in the future, the more changes the class need to apply.

Reference: DotNetTricks. SOLID Design Principles Explained - C#. Retrieved from https://www.dotnettricks.com/learn/designpatterns/solid-design-principles-explained-using-csharp on September 20, 2023