Open-Closed Principle
The Open/Closed Principle says, "A software module/class is open for extension and closed for modification." In this implementation, I created a new class AuthorCorporate to handle a book's author if it is a company who published it.

Note: This example is influenced by the MARC Code 110/710 used by librarians to encode book details to an Integrated Library System (ILS).

Resources used:
https://www.c-sharpcorner.com/UploadFile/damubetha/solid-principles-in-C-Sharp/
https://www.oclc.org/bibformats/en/1xx/110.html
https://www.oclc.org/bibformats/en/7xx/710.html