Liskov Substitution Principle
The Liskov Substitution Principle says that the object of a derived class should be able to replace an object of the base class without bringing any errors in the system or modifying the behavior of the base class. That means child class objects should be able to replace parent class objects without compromising application integrity.

Source: https://dotnettutorials.net/lesson/liskov-substitution-principle/